+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Expose gRPC metrics (#584)
+
 ## [0.6.1] - 2022-03-01
 
 ### Added